Altered login service to use AuthenticatedUser type

+++ web/src/services/contestService.ts
@@ -1,3 +1,4 @@
+import { getAccessToken } from '@/utils/authUtility';
 import { Contest, CreateContestRequest, Event } from '@raffle-tracker/dto';
 
 const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
@@ -7,6 +8,11 @@ export async function addContest(
   contestNumber: number,
   numberOfHorses: number
 ): Promise<Event> {
+  const token = await getAccessToken();
+  if (!token) {
+    throw new Error('Not authenticated');
+  }
+
   const createContestRequest: CreateContestRequest = {
     eventId,
     contestNumber,
@@ -16,6 +22,7 @@ export async function addContest(
     method: 'POST',
     body: JSON.stringify(createContestRequest),
     headers: {
+      Authorization: `Bearer ${token}`,
       'Content-Type': 'application/json',
     },
   });
@@ -30,8 +37,16 @@ export async function addContest(
 }
 
 export async function deleteContest(id: number): Promise<void> {
+  const token = await getAccessToken();
+  if (!token) {
+    throw new Error('Not authenticated');
+  }
+
   const res = await fetch(`${API_BASE_URL}/contests/${id}`, {
     method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   });
   if (!res.ok) {
     const errorData = await res.json();
@@ -43,8 +58,14 @@ export async function getContest(
   id: number,
   includeChildren: boolean
 ): Promise<Contest> {
+  const token = await getAccessToken();
+  if (!token) {
+    throw new Error('Not authenticated');
+  }
+
   const res = await fetch(
-    `${API_BASE_URL}/contests/${id}${includeChildren ? '?includeChildren=true' : ''}`
+    `${API_BASE_URL}/contests/${id}${includeChildren ? '?includeChildren=true' : ''}`,
+    { headers: { Authorization: `Bearer ${token}` } }
   );
 
   if (!res.ok) {
@@ -59,7 +80,14 @@ export async function getContest(
 export async function getValidContestsByEvent(
   eventId: number
 ): Promise<Contest[]> {
-  const res = await fetch(`${API_BASE_URL}/events/${eventId}/races`);
+  const token = await getAccessToken();
+  if (!token) {
+    throw new Error('Not authenticated');
+  }
+
+  const res = await fetch(`${API_BASE_URL}/events/${eventId}/races`, {
+    headers: { Authorization: `Bearer ${token}` },
+  });
   if (!res.ok) {
     const errorData = await res.json();
     throw new Error(errorData.message || 'Failed to fetch contests');
